Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,8 +4,8 @@
 int main() {
     float r;
     scanf("%f",&r);
-    float are= (3.14*(r*r));
-    printf("Area: %f",are);
+    int are= (3.14*(r*r));
+    printf("Area: %d",are);
 
     return 0;
 }
